requirements.txt
# Core ML frameworks
torch>=2.0.0
torchvision>=0.15.0

# Transformers for LLM and vision models
transformers>=4.30.0
accelerate>=0.20.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.65.0
numpy>=1.24.0
pandas>=2.0.0

# Metrics and logging
scikit-learn>=1.3.0
tensorboard>=2.13.0

# Optional: for specific vision encoders
timm>=0.9.0

# Development
pytest>=7.4.0
black>=23.0.0
